# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Connect platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  uid: string;
  username: string;
  displayName: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  isOnline: boolean;
  inviteLink: string;
  createdAt: string;
}

export interface UserSummary {
  id: number;
  username: string;
  displayName: string;
  /** @nullable */
  avatarUrl?: string | null;
  isOnline: boolean;
}

export interface UserUpdate {
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface UsernameCheck {
  username: string;
  available: boolean;
}

export type ConnectionStatus = typeof ConnectionStatus[keyof typeof ConnectionStatus];


export const ConnectionStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
  blocked: 'blocked',
} as const;

export interface Connection {
  id: number;
  status: ConnectionStatus;
  user: UserSummary;
  createdAt: string;
}

export interface ConnectionInput {
  username: string;
}

export interface ConnectionRequests {
  received: Connection[];
  sent: Connection[];
}

/**
 * @nullable
 */
export type ConversationLastMessage = {
  text: string;
  priority: string;
  sentAt: string;
} | null;

export interface Conversation {
  id: number;
  participant: UserSummary;
  /** @nullable */
  lastMessage?: ConversationLastMessage;
  unreadCount: number;
  isPinned: boolean;
  /** @nullable */
  wallpaper?: string | null;
}

export type MessagePriority = typeof MessagePriority[keyof typeof MessagePriority];


export const MessagePriority = {
  normal: 'normal',
  good_news: 'good_news',
  important: 'important',
  urgent: 'urgent',
  emergency: 'emergency',
} as const;

export interface Message {
  id: number;
  conversationId: number;
  senderId: number;
  /** @nullable */
  text?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  priority: MessagePriority;
  isRead: boolean;
  /** @nullable */
  replyToId?: number | null;
  sentAt: string;
}

export type MessageInputPriority = typeof MessageInputPriority[keyof typeof MessageInputPriority];


export const MessageInputPriority = {
  normal: 'normal',
  good_news: 'good_news',
  important: 'important',
  urgent: 'urgent',
  emergency: 'emergency',
} as const;

export interface MessageInput {
  text?: string;
  imageUrl?: string;
  priority: MessageInputPriority;
  replyToId?: number;
}

export interface EmergencyAlert {
  id: number;
  fromUser: UserSummary;
  message: string;
  sentAt: string;
  isResolved: boolean;
}

export interface EmergencyAlertInput {
  message: string;
  targetUsername: string;
}

export interface Theme {
  id: string;
  name: string;
  previewGradient: string;
  isActive: boolean;
}

export interface Settings {
  activeTheme: string;
  autoLogoutEnabled: boolean;
  autoLogoutMinutes: number;
  notificationsEnabled: boolean;
  emergencyEmailEnabled: boolean;
  emergencySmsEnabled: boolean;
}

export interface SettingsUpdate {
  activeTheme?: string;
  autoLogoutEnabled?: boolean;
  autoLogoutMinutes?: number;
  notificationsEnabled?: boolean;
  emergencyEmailEnabled?: boolean;
  emergencySmsEnabled?: boolean;
}

export type DeviceDeviceType = typeof DeviceDeviceType[keyof typeof DeviceDeviceType];


export const DeviceDeviceType = {
  android: 'android',
  laptop: 'laptop',
  desktop: 'desktop',
  tablet: 'tablet',
  phone: 'phone',
} as const;

export interface Device {
  id: number;
  deviceType: DeviceDeviceType;
  deviceName: string;
  browser: string;
  lastSeen: string;
  isCurrent: boolean;
}

export type CheckUsernameParams = {
username: string;
};

export type SearchUsersParams = {
q: string;
};

export type ListMessagesParams = {
limit?: number;
before?: string;
};

export type LogoutAllDevices200 = {
  success: boolean;
};