# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Connect platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User profile operations
  - name: connections
    description: Connection management
  - name: conversations
    description: Chat conversations
  - name: messages
    description: Messaging
  - name: themes
    description: Themes and settings
  - name: devices
    description: Device management
  - name: emergency
    description: Emergency alerts
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users/check-username:
    get:
      operationId: checkUsername
      tags: [users]
      summary: Check if username is available
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Availability result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsernameCheck"

  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"

  /connections:
    get:
      operationId: getConnections
      tags: [connections]
      summary: Get all connections
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
    post:
      operationId: sendConnectionRequest
      tags: [connections]
      summary: Send a connection request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionInput"
      responses:
        "201":
          description: Connection request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"

  /connections/requests:
    get:
      operationId: getConnectionRequests
      tags: [connections]
      summary: Get pending connection requests
      responses:
        "200":
          description: Pending requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionRequests"

  /connections/{connectionId}/accept:
    post:
      operationId: acceptConnection
      tags: [connections]
      summary: Accept a connection request
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"

  /connections/{connectionId}/reject:
    post:
      operationId: rejectConnection
      tags: [connections]
      summary: Reject a connection request
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"

  /conversations:
    get:
      operationId: getConversations
      tags: [conversations]
      summary: Get all conversations
      responses:
        "200":
          description: Conversation list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"

  /conversations/{conversationId}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a single conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{conversationId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  /emergency/alerts:
    get:
      operationId: getEmergencyAlerts
      tags: [emergency]
      summary: Get emergency alerts
      responses:
        "200":
          description: Emergency alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmergencyAlert"
    post:
      operationId: sendEmergencyAlert
      tags: [emergency]
      summary: Send an emergency alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyAlertInput"
      responses:
        "201":
          description: Alert sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyAlert"

  /themes:
    get:
      operationId: getThemes
      tags: [themes]
      summary: Get available themes
      responses:
        "200":
          description: Themes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Theme"

  /settings:
    get:
      operationId: getSettings
      tags: [themes]
      summary: Get user settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [themes]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

  /devices:
    get:
      operationId: getDevices
      tags: [devices]
      summary: Get connected devices
      responses:
        "200":
          description: Devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"

  /devices/{deviceId}/logout:
    post:
      operationId: logoutDevice
      tags: [devices]
      summary: Logout a specific device
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Device logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"

  /devices/logout-all:
    post:
      operationId: logoutAllDevices
      tags: [devices]
      summary: Logout all devices
      responses:
        "200":
          description: All devices logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    User:
      type: object
      properties:
        id:
          type: integer
        uid:
          type: string
        username:
          type: string
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isOnline:
          type: boolean
        inviteLink:
          type: string
        createdAt:
          type: string
      required: [id, uid, username, displayName, isOnline, inviteLink, createdAt]

    UserSummary:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isOnline:
          type: boolean
      required: [id, username, displayName, isOnline]

    UserUpdate:
      type: object
      properties:
        displayName:
          type: string
        bio:
          type: string
        avatarUrl:
          type: string

    UsernameCheck:
      type: object
      properties:
        username:
          type: string
        available:
          type: boolean
      required: [username, available]

    Connection:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected, blocked]
        user:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
      required: [id, status, user, createdAt]

    ConnectionInput:
      type: object
      properties:
        username:
          type: string
      required: [username]

    ConnectionRequests:
      type: object
      properties:
        received:
          type: array
          items:
            $ref: "#/components/schemas/Connection"
        sent:
          type: array
          items:
            $ref: "#/components/schemas/Connection"
      required: [received, sent]

    Conversation:
      type: object
      properties:
        id:
          type: integer
        participant:
          $ref: "#/components/schemas/UserSummary"
        lastMessage:
          type: ["object", "null"]
          properties:
            text:
              type: string
            priority:
              type: string
            sentAt:
              type: string
          required: [text, priority, sentAt]
        unreadCount:
          type: integer
        isPinned:
          type: boolean
        wallpaper:
          type: ["string", "null"]
      required: [id, participant, unreadCount, isPinned]

    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        senderId:
          type: integer
        text:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        priority:
          type: string
          enum: [normal, good_news, important, urgent, emergency]
        isRead:
          type: boolean
        replyToId:
          type: ["integer", "null"]
        sentAt:
          type: string
      required: [id, conversationId, senderId, priority, isRead, sentAt]

    MessageInput:
      type: object
      properties:
        text:
          type: string
        imageUrl:
          type: string
        priority:
          type: string
          enum: [normal, good_news, important, urgent, emergency]
        replyToId:
          type: integer
      required: [priority]

    EmergencyAlert:
      type: object
      properties:
        id:
          type: integer
        fromUser:
          $ref: "#/components/schemas/UserSummary"
        message:
          type: string
        sentAt:
          type: string
        isResolved:
          type: boolean
      required: [id, fromUser, message, sentAt, isResolved]

    EmergencyAlertInput:
      type: object
      properties:
        message:
          type: string
        targetUsername:
          type: string
      required: [message, targetUsername]

    Theme:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        previewGradient:
          type: string
        isActive:
          type: boolean
      required: [id, name, previewGradient, isActive]

    Settings:
      type: object
      properties:
        activeTheme:
          type: string
        autoLogoutEnabled:
          type: boolean
        autoLogoutMinutes:
          type: integer
        notificationsEnabled:
          type: boolean
        emergencyEmailEnabled:
          type: boolean
        emergencySmsEnabled:
          type: boolean
      required: [activeTheme, autoLogoutEnabled, autoLogoutMinutes, notificationsEnabled, emergencyEmailEnabled, emergencySmsEnabled]

    SettingsUpdate:
      type: object
      properties:
        activeTheme:
          type: string
        autoLogoutEnabled:
          type: boolean
        autoLogoutMinutes:
          type: integer
        notificationsEnabled:
          type: boolean
        emergencyEmailEnabled:
          type: boolean
        emergencySmsEnabled:
          type: boolean

    Device:
      type: object
      properties:
        id:
          type: integer
        deviceType:
          type: string
          enum: [android, laptop, desktop, tablet, phone]
        deviceName:
          type: string
        browser:
          type: string
        lastSeen:
          type: string
        isCurrent:
          type: boolean
      required: [id, deviceType, deviceName, browser, lastSeen, isCurrent]